// src/app/shared/interfaces/cpps.interface.ts
export interface Cpps {
  cpps: Cpp[];
}
export interface Cpp {
        'id_cpp': number;
        'id_mem': number;
        'reg_cpp_tip_id': number;
        'reg_cpp_id': number;
        'grad_prof_cpp_id': number;
        'date_start': string;
        'date_end': string;
        'emitent': string;
        'act_serie': string;
        'act_numar': string;
        'act_data': string;
        'act_descriere': string;
        'obs': string;
        'updated': string;
        'ro': string;
}